+++ src/ui/__tests__/DrivingHUD.test.tsx
@@ -8,11 +8,11 @@ jest.mock('react-native', () => ({
   Text: 'Text',
   StyleSheet: { create: () => ({}), flatten: () => ({}) },
 }));
-jest.mock('../premium/subscription', () => ({
+jest.mock('../../premium/subscription', () => ({
   usePremium: () => ({ isPremium: true }),
 }));
 
-import DrivingHUD from '../components/DrivingHUD';
+import DrivingHUD from '../DrivingHUD';
 
 describe('DrivingHUD', () => {
   it('displays provided props', () => {
